Refactor handler methods
- MainWindowUI.createSearchWindow
- MainWindowUI.findInProjectReuseLastWindow
- MainWindowUI.getTrimmedSelectedTextInMainWindow
- MainWindowUI.projectExit
- MainWindowUI.projectRestart
- MainWindowUI.prepareForExit
